requirements.txt
azure-storage-blob>=12.19.0